Adding exception handler and the exception classes

+++ Api-Java/landpage/src/main/java/br/com/jonsons/landpage/service/EnderecoService.java
@@ -7,22 +7,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.jonsons.landpage.domain.Enderecos;
+import br.com.jonsons.landpage.exceptions.EnderecoNotFoundException;
 import br.com.jonsons.landpage.repository.EnderecoRepository;
-
 
 @Service
 public class EnderecoService {
 	@Autowired 
 	EnderecoRepository enderecoRepository;
 	
-	public List<Enderecos> getAll(){
+	public List<Enderecos> getAll() {
 		return enderecoRepository.findAll();
 	}
 	
-	public Optional<Enderecos> getById(Long codigo) {
+	public Optional<Enderecos> getById(Long codigo) throws EnderecoNotFoundException {
 		Optional<Enderecos> endereco = enderecoRepository.findById(codigo);
 		if(!endereco.isPresent()) {
-			return null;
+			throw new EnderecoNotFoundException();
 		}
 		return endereco;
 	}
@@ -30,12 +30,13 @@ public class EnderecoService {
 	public Enderecos post(Enderecos endereco) {
 		return enderecoRepository.save(endereco);
 	}
-	
-	public Enderecos update(Long codigo, Enderecos endereco) {
+
+	public Enderecos update(Long codigo, Enderecos endereco) throws EnderecoNotFoundException {
 		 Optional<Enderecos> enderecoAtualizado = enderecoRepository.findById(codigo);
 		 if(!enderecoAtualizado.isPresent()) {
-			 return null;
+			 throw new EnderecoNotFoundException();
 		 }
+		 
 		 Enderecos end = enderecoAtualizado.get();
 		 if(endereco.getBairro() != null) {
 			 end.setBairro(endereco.getBairro());
@@ -65,8 +66,11 @@ public class EnderecoService {
 		 return end;
 		}
 	
-		public void delete(Long codigo) {
+		public void delete(Long codigo) throws EnderecoNotFoundException {
 			Optional<Enderecos> endereco = enderecoRepository.findById(codigo);
+			if(!endereco.isPresent()) {
+				throw new EnderecoNotFoundException();
+			}
 			enderecoRepository.delete(endereco.get());
 		}
 }
